move some defines to codex_defines.hpp
